+++ tests/test-node.js
@@ -3,12 +3,15 @@
   factory();
 }((function () { 'use strict';
 
+  /* eslint-disable node/no-unsupported-features/es-syntax,
+    node/no-unsupported-features/es-builtins */
   const write = (...msgs) => {
     if (typeof document !== 'undefined') {
       document.body.append(
         ...msgs, ...Array.from({length: 2}, () => document.createElement('br'))
       );
     } else {
+      // eslint-disable-next-line no-console
       console.log(...msgs);
     }
   };
@@ -32,17 +35,39 @@
     }
   };
 
+  /* eslint-disable node/no-unsupported-features/es-syntax */
+
+  /**
+  * @callback SimpleJSONCallback
+  * @param {JSON} json
+  * @returns {void}
+  */
+
+  /**
+  * @callback SimpleJSONErrback
+  * @param {Error} err
+  * @param {string|string[]} jsonURL
+  * @returns {void}
+  */
+
+  /**
+   * @param {string|string[]} jsonURL
+   * @param {SimpleJSONCallback} cb
+   * @param {SimpleJSONErrback} errBack
+   * @returns {Promise<JSON>}
+   */
   async function getJSON (jsonURL, cb, errBack) {
     try {
       if (Array.isArray(jsonURL)) {
         const arrResult = await Promise.all(jsonURL.map((url) => getJSON(url)));
         if (cb) {
-          // eslint-disable-next-line callback-return, standard/no-callback-literal
+          // eslint-disable-next-line node/callback-return, standard/no-callback-literal, promise/prefer-await-to-callbacks
           cb(...arrResult);
         }
         return arrResult;
       }
       const result = await fetch(jsonURL).then((r) => r.json());
+      // eslint-disable-next-line promise/prefer-await-to-callbacks
       return typeof cb === 'function' ? cb(result) : result;
     } catch (e) {
       e.message += ` (File: ${jsonURL})`;
@@ -53,12 +78,13 @@
     }
   }
 
-  /* eslint-env node */
+  /* eslint-disable node/no-unsupported-features/es-syntax */
 
   if (typeof fetch === 'undefined') {
     global.fetch = (jsonURL) => {
+      // eslint-disable-next-line promise/avoid-new
       return new Promise((resolve, reject) => {
-        // eslint-disable-next-line global-require
+        // eslint-disable-next-line node/global-require, no-shadow
         const XMLHttpRequest = require('local-xmlhttprequest')({
           basePath: __dirname
         }); // Don't change to an import as won't resolve for browser testing
@@ -85,23 +111,27 @@
     };
   }
 
-  /* eslint-disable handle-callback-err */
+  /* eslint-disable node/no-unsupported-features/es-syntax */
 
   const getJSON$1 = typeof module === 'undefined' ? getJSON : getJSON;
 
   // eslint-disable-next-line promise/always-return
   getJSON$1('test.json').then((result) => {
     assert.equals(5, result.key, 'Retrieve JSON result value - single string URL (normal promise)');
+    // eslint-disable-next-line promise/prefer-await-to-callbacks
   }).catch((err) => {
+    // eslint-disable-next-line no-console
+    console.log(err);
     assert.true(false, `Shouldn't get here`);
+    return true;
   });
 
   (async () => {
   const result = await getJSON$1('test.json');
   assert.equals(5, result.key, 'Retrieve JSON result value - single string URL (await)');
 
-  await getJSON$1('test.json', (result) => {
-    assert.equals(5, result.key, 'Retrieve JSON result value - single string URL (callback)');
+  await getJSON$1('test.json', (reslt) => {
+    assert.equals(5, reslt.key, 'Retrieve JSON result value - single string URL (callback)');
   });
 
   await getJSON$1('test-nonexisting.json', () => {