+++ tests/test-utils.js
@@ -1,9 +1,12 @@
+/* eslint-disable node/no-unsupported-features/es-syntax,
+  node/no-unsupported-features/es-builtins */
 const write = (...msgs) => {
   if (typeof document !== 'undefined') {
     document.body.append(
       ...msgs, ...Array.from({length: 2}, () => document.createElement('br'))
     );
   } else {
+    // eslint-disable-next-line no-console
     console.log(...msgs);
   }
 };